requirements.txt
streamlit
pandas
plotly-express
matplotlib
seaborn
joblib
numpy
scikit-learn
altair
streamlit-option-menu
googletrans==4.0.0-rc1